DA::Client renaming refactor from RunLoop::XCUITest
